fix: change composable name [WTEL-6752] (https://webitel.atlassian.net/browse/WTEL-6752)
