+++ src/ui/modules/queue-section/modules/chat-queue/components/manual-queue/manual-queue-container.vue
@@ -19,7 +19,7 @@
 import { computed } from 'vue';
 import { useStore } from 'vuex';
 
-import { useLoadingState } from '../../../../../../composables/useLoadingState';
+import { useLoader } from '../../../../../../composables/useLoader';
 import TaskQueueContainer from '../../../_shared/components/task-queue-container.vue';
 import ManualPreview from './manual-queue-preview.vue';
 
@@ -32,14 +32,14 @@ const props = defineProps({
 
 const store = useStore();
 
-const { showLoader, withLoading } = useLoadingState();
+const { showLoader, runWithLoader } = useLoader();
 
 console.info(store.state.features.chat.manual.manualList);
 
 const manualList = computed(() => store.state.features.chat.manual.manualList);
 
 function acceptTask(task) {
-	return withLoading(task.id, () =>
+	return runWithLoader(task.id, () =>
 		store.dispatch('features/chat/manual/ACCEPT_TASK', task),
 	);
 }
