+++ src/ui/modules/work-section/modules/call/components/call-contacts/contacts/contacts-container.vue
@@ -49,7 +49,7 @@ import { useStore } from 'vuex';
 
 import SearchMode from '../../../../../../../../app/api/agent-workspace/endpoints/contacts/enums/SearchMode.enum';
 import useInfiniteScroll from '../../../../../../../../app/composables/useInfiniteScroll';
-import { useLoadingState } from '../../../../../../../composables/useLoadingState';
+import { useLoader } from '../../../../../../../composables/useLoader';
 import { useUserinfoStore } from '../../../../../../userinfo/userinfoStore';
 import LookupItemContainer from '../../../../_shared/components/lookup-item-container/lookup-item-container.vue';
 import EmptySearch from '../../../../_shared/components/workspace-empty-search/components/empty-search.vue';
@@ -65,7 +65,7 @@ const props = defineProps({
 const { t } = useI18n();
 const store = useStore();
 const userinfoStore = useUserinfoStore();
-const { showLoader, withLoading } = useLoadingState();
+const { showLoader, runWithLoader } = useLoader();
 
 const filterQuery = ref(SearchMode.NAME);
 const contactsLabelsConfiguration = ref([]);
@@ -120,7 +120,9 @@ const fetchFn = async (params) => {
 };
 
 const makeCall = (item) => {
-	withLoading(item.contactId, () => store.dispatch('features/call/CALL', item));
+	runWithLoader(item.contactId, () =>
+		store.dispatch('features/call/CALL', item),
+	);
 };
 
 const changeMode = ({ value }) => {
