+++ src/ui/modules/work-section/modules/call/components/call-contacts/users/users-container.vue
@@ -30,7 +30,7 @@ import { computed } from 'vue';
 import { useStore } from 'vuex';
 
 import useInfiniteScroll from '../../../../../../../../app/composables/useInfiniteScroll';
-import { useLoadingState } from '../../../../../../../composables/useLoadingState';
+import { useLoader } from '../../../../../../../composables/useLoader';
 import { useUserinfoStore } from '../../../../../../userinfo/userinfoStore';
 import UserLookupItem from '../../../../_shared/components/lookup-item/user-lookup-item.vue';
 import LookupItemContainer from '../../../../_shared/components/lookup-item-container/lookup-item-container.vue';
@@ -46,7 +46,7 @@ const props = defineProps({
 const store = useStore();
 
 const userinfoStore = useUserinfoStore();
-const { showLoader, withLoading } = useLoadingState();
+const { showLoader, runWithLoader } = useLoader();
 
 const { userId } = storeToRefs(userinfoStore);
 
@@ -72,7 +72,7 @@ const { dataList, isLoading, dataSearch, handleIntersect, resetData } =
 	});
 
 const makeCall = (item) => {
-	withLoading(item.number, () => store.dispatch('features/call/CALL', item));
+	runWithLoader(item.number, () => store.dispatch('features/call/CALL', item));
 };
 </script>
